+++ app/page.tsx
@@ -60,9 +60,9 @@ export default function Home() {
           Playground
         </Link>
       </div>
-      <div className="flex flex-row items-start sm:gap-2 gap-0.5 text-muted-foreground text-md mt-7 mb-12 font-code text-base font-medium">
+      <div className="flex flex-row items-start sm:gap-2 gap-0.5 text-muted-foreground text-md mt-10 mb-12 font-code text-base font-medium">
         <TerminalSquareIcon className="w-5 h-5 mr-1 mt-0.5" />
-        {"npx create_docu"}
+        {"npx @docubook/create@latest"}
       </div>
       <NpxTerminal/>
     </div>